Pulling

+++ Swiftly/DataStructures.swift
@@ -28,7 +28,7 @@ struct User: Codable{
     var userName: String
     var streakCount: Int = 0
     var currentLevel: Int = 0
-    var totalScore:Int = 0
+    var totalScore: Double = 0
     var chapterScores:[Int]
     
     var lastLogIn: TimeInterval
